Reimplement verifyAllCombinations as variadic function
Fixes #33

The parameter order needed to change, since parameter pack types
can only be deduced if the pack is the last parameter.

For backwards compatibility overloads with the reporter at the end
are provided for up to 9 inputs (as was the case before).

Also put the functions in a namespace instead of a class.