t Fix compilation failure in the Xcode CI builds that use unity.
See comments in #89 for the workaround to this.
For now, I've excluded CombinationTests.cpp from the unity builds,
so it will be compiled on its own.

I have confirmed the same problem occurs in XCode with unity builds
and the single header - so this will need investigation at some point.