F Include version-number info in development code. (#40)
A later step will update the stored version during releases.

This adds a new header, ApprovalTestsVersion.h, which contains the
current release version info.

The header Blocker.h seems to always be include first in the
single header, at least going back to v.3.3.0, so by including
this new header in BLocker.h, we ensure that the information
from this file will always appear at the top of the generated
single-header file.