F!! Add support for SublimeMerge
ApprovalTests can now use SublimeMerge on Linux, macOS, and Windows.
On Linux, there is additional support for the SublimeMerge Snap, Flatpak, and tarball installations.
In the case of multiple installation on Linux, the one chosen is determined as follows:
1. Snap
2. Flatpak
3. Any tool on the path named "smerge"
4. The installed tarball

Please note that while the Flatpak package is official, the Snap package is unofficial at this time.
The markdown snippets and single-header file will need to be regenerated.
A refactor of how merge tools are found in the future should reduce much of the boilerplate introduced here.