F!! Add namer that allows users to full customization
This namer can be used to completely cusomize where approved and
received files will be stored. This allows to e.g. make the path
dependend on the run time location of the test binary instead of
the source location for cases where the machine running the tests
might not even have the source files checked out (e.g. when cross-
compiling for ARM on a x86 machine).