d Make table column widths consistent, before making them narrower.
When GettingStarted.md is converted to ReStructuredText by pandoc,
if the columns are too wide for pandoc's liking, it makes a huge
mess of the table, by splitting lots of words. They are then
shown in sphinx with spaces in the middle of the words.

See https://github.com/jgm/pandoc/issues/2612

The only way I've found to work around this is to remove any
hyperlinks from the table, and shorten some words, which
I will do in later stages.