e   Remove need to manually delete branch in conan-center-index repo
This is caused by running a prepare_*_release.sh script
and then not proceeding with the release...

And then, some time later, repeating that step, during which
time more commits have been made in that conan repo.

This was a major annoyance in doing our releases, as the prepare
step failed due to a branch already existing with the name of the
new release, and having different content.

Now a prompt is issued, to ask if the pre-existing branch should
be deleted.